Move region test to plan-selector
